+++ src/main/java/com/google/codeu/servlets/MessageServlet.java
@@ -93,7 +93,7 @@ public class MessageServlet extends HttpServlet {
 }
 
 
-  private float getSentimentScore(String text) throws IOException {
+  private double getSentimentScore(String text) throws IOException {
   Document doc = Document.newBuilder()
       .setContent(text).setType(Type.PLAIN_TEXT).build();
 
@@ -101,7 +101,7 @@ public class MessageServlet extends HttpServlet {
   Sentiment sentiment = languageService.analyzeSentiment(doc).getDocumentSentiment();
   languageService.close();
 
-  return sentiment.getScore();
+  return (double) sentiment.getScore();
 }
 
   /** Stores a new {@link Message}. */
@@ -123,7 +123,7 @@ public class MessageServlet extends HttpServlet {
     String regex = "(https?://\\S+\\.(png|jpg|gif))";
     String replacement = "<img src=\"$1\" />";
     String textWithImagesReplaced = userText.replaceAll(regex, replacement);
-    float sentimentScore = getSentimentScore(textWithImagesReplaced);
+    double sentimentScore = getSentimentScore(textWithImagesReplaced);
     Message message = new Message(user, textWithImagesReplaced, recipient, sentimentScore);
     datastore.storeMessage(message);
